src/FizzyPrinter.java
public class FizzyPrinter {

    public String printFizzy(int n, boolean upperCase) {
        String result = "";

        if (n % 3 == 0 && n % 5 == 0 && n % 7 == 0) {
            result = "Fizzbuzzbang";
        } else if (n % 3 == 0 && n % 5 == 0) {
            result = "Fizzbuzz";
        } else if (n % 3 == 0) {
            result = "Fizz";
        } else if (n % 5 == 0) {
            result = "Buzz";
        } else if (n % 7 == 0) {
            result = "Bang";
        } else {
            result = "Boom";
        }

        if (upperCase) {
            result = result.toUpperCase();
        }

        return result;
    }
}
